Refactor Frontend Service Layer
